Ads + Python/cloudflare_turnstile_request/pyproject.toml
[project]
name = "cloudflare-turnstile-request"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "2captcha-python>=1.5.1",
    "better-proxy>=1.3.1",
    "fake-useragent>=2.2.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
